[videoMetadata] add metadataRepo
