--- extraction/src/main/java/org/hstack/vmeta/extraction/Model/Metadata/CategoryType.java
@@ -1,29 +1,0 @@
-package org.hstack.vmeta.extraction.Model.Metadata;
-
-public enum CategoryType {
-
-    POLITICS,       // 정치, 행정
-    ECONOMY,        // 경제
-    EDUCATION,      // 교육
-    MEDIA,          // 미디어
-    GEOGRAPHY,      // 지리
-    SPORTS,         // 스포츠
-    SCIENCE,        // 과학
-    LAW,            // 법학
-    MEDICINE,       // 의학
-    FOOD,           // 요리
-    ARCHITECTURE,   // 건축
-    MUSIC,          // 음악
-    CULTURE,        // 문화
-    TRANSPORT,      // 교통
-    ART,            // 예술 (패션 포함)
-    LINGUISTICS,    // 언어
-    PHILOSOPHY,     // 철학 (종교 포함)
-    HISTORY,        // 역사
-    NATURE,         // 자연
-    MILITARY,       // 군사
-    TRAVEL,         // 여행
-    PEOPLE,         // 인물
-    IT              // IT
-
-}
