--- extraction/src/main/java/org/hstack/vmeta/extraction/Model/Metadata/ContentWithTime.java
@@ -1,17 +1,0 @@
-package org.hstack.vmeta.extraction.Model.Metadata;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.sql.Time;
-
-@Getter
-@Setter
-@Builder
-public class ContentWithTime {
-
-    private Time time;
-    private String content;
-
-}
