--- extraction/src/main/java/org/hstack/vmeta/extraction/Model/Metadata/Narrative.java
@@ -1,7 +1,0 @@
-package org.hstack.vmeta.extraction.Model.Metadata;
-
-public enum Narrative {
-    
-    APPLICATION,    // 실습 같이 현장감 있는 형식
-    DESCRIPTION     // 특정 자료가 제공되는 형식
-}
