--- extraction/src/main/java/org/hstack/vmeta/extraction/Model/Metadata/Presentation.java
@@ -1,8 +1,0 @@
-package org.hstack.vmeta.extraction.Model.Metadata;
-
-public enum Presentation {
-
-    STATIC,
-    DYNAMIC
-
-}
