--- extraction/src/main/java/org/hstack/vmeta/extraction/Model/Metadata/VideoFrame.java
@@ -1,6 +1,0 @@
-package org.hstack.vmeta.extraction.Model.Metadata;
-
-public enum VideoFrame {
-    FPS_24,
-    FPS_30
-}
