--- extraction/src/main/java/org/hstack/vmeta/extraction/Model/Metadata/VideoType.java
@@ -1,6 +1,0 @@
-package org.hstack.vmeta.extraction.Model.Metadata;
-
-public enum VideoType {
-    MP4,
-    AVI
-}
