+++ videoMetadata/gradle/wrapper/gradle-wrapper.properties
