+++ videoMetadata/settings.gradle
