+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/VideoMetadataApplication.java
