+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/category/CategoryAttributeConverter.java
@@ -1,0 +1,30 @@
+package org.hstack.vmeta.videoMetadata.metadata.category;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.persistence.AttributeConverter;
+
+import java.util.List;
+
+public class CategoryAttributeConverter implements AttributeConverter<List<Category>, String> {
+
+    private final ObjectMapper mapper = new ObjectMapper();
+
+    @Override
+    public String convertToDatabaseColumn(List<Category> categoryList) {
+        try {
+            return mapper.writeValueAsString(categoryList);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public List<Category> convertToEntityAttribute(String s) {
+        try {
+            return mapper.readValue(s, List.class);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
