+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/index/Index.java
@@ -1,0 +1,14 @@
+package org.hstack.vmeta.videoMetadata.metadata.index;
+
+import jakarta.persistence.Embeddable;
+
+import java.sql.Time;
+
+@Embeddable
+public class Index {
+
+    private Time time;
+    private String content;
+    private boolean expose;
+    private boolean autocreated;
+}
