+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/keyword/Keyword.java
@@ -1,0 +1,11 @@
+package org.hstack.vmeta.videoMetadata.metadata.keyword;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+public class Keyword {
+    private String keyword;
+    private float perc;
+    private boolean expose;
+    private boolean autocreated;
+}
