+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/narrative/Narrative.java
@@ -1,0 +1,6 @@
+package org.hstack.vmeta.videoMetadata.metadata.narrative;
+
+public enum Narrative {
+    APPLICATION,    // 실습 같이 현장감 있는 형식
+    DESCRIPTION     // 특정 자료가 제공되는 형식
+}
