+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/presentation/Presentation.java
@@ -1,0 +1,6 @@
+package org.hstack.vmeta.videoMetadata.metadata.presentation;
+
+public enum Presentation {
+    STATIC,
+    DYNAMIC
+}
