+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/script/Script.java
@@ -1,0 +1,13 @@
+package org.hstack.vmeta.videoMetadata.metadata.script;
+
+import jakarta.persistence.Embeddable;
+
+import java.sql.Time;
+
+@Embeddable
+public class Script {
+
+    private Time time;
+    private String content;
+
+}
