+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/videoFrame/VideoFrame.java
@@ -1,0 +1,6 @@
+package org.hstack.vmeta.videoMetadata.metadata.videoFrame;
+
+public enum VideoFrame {
+    FPS24,
+    FPS30
+}
