+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/metadata/videoType/VideoType.java
@@ -1,0 +1,6 @@
+package org.hstack.vmeta.videoMetadata.metadata.videoType;
+
+public enum VideoType {
+    MP4,
+    AVI
+}
