+++ videoMetadata/src/main/java/org/hstack/vmeta/videoMetadata/video/Video.java
@@ -1,0 +1,25 @@
+package org.hstack.vmeta.videoMetadata.video;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@Entity
+@Table(name="video")
+public class Video {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long Id;
+
+    @Column
+    private String title;
+
+    @Column
+    private String path;
+
+    @Column
+    private String uploaderName;
+}
