+++ videoMetadata/src/test/java/org/hstack/vmeta/videoMetadata/VideoMetadataApplicationTests.java
